+++ Core/DTRichTextEditor.h
@@ -10,12 +10,18 @@
 // specialized RTE subclasses
 #import "DTTextRange.h"
 #import "DTTextPosition.h"
+#import "DTTextSelectionRect.h"
 
+// UI
 #import "DTRichTextEditorContentView.h"
 
 #import "DTRichTextEditorView.h"
 #import "DTRichTextEditorView+DTCoreText.h"
 #import "DTRichTextEditorView+Manipulation.h"
+#import "DTRichTextEditorView+Dictation.h"
+
+#import "DTTextSelectionView.h"
+#import "DTDictationPlaceholderView.h"
 
 // time bomb set by integration server for nightly demo build
 #ifdef TIMEBOMB