+++ Documentation/Changelog.markdown
@@ -2,6 +2,7 @@ Version 1.1.4
 - FIXED: horizontal flickering when moving round loupe over text
 - FIXED: cursor does not stop blinking during selection
 - UPDATED: DTCoreText to 1.2.1
+- FIXED: content size problem caused by DTCoreText change
 - CHANGED: renamed contentView to attributedTextContentView to avoid possible conflict with internal ivar of UIScrollView
 - CHANGED: replaced semaphore-based sync with dispatch_queue
 -