+++ Documentation/Changelog.markdown
@@ -1,6 +1,8 @@
 Version 1.1.4
 - FIXED: horizontal flickering when moving round loupe over text
 - FIXED: cursor does not stop blinking during selection
+- FIXED: see-through mode of loupe used when touch leaves visible content area
+- CHANGED: restrict loupe towards bottom so that it does not go under keyboard
 - UPDATED: DTCoreText to 1.2.1
 - FIXED: content size problem caused by DTCoreText change
 - CHANGED: renamed contentView to attributedTextContentView to avoid possible conflict with internal ivar of UIScrollView